chore: move interfaces to index.d.ts
